Use fulfill or reject explicitly when resolving request promise
Calling resolve() on a resolver cannot have both args as nil. First check if response vaildates, if not - reject with error; then check if response has an error, and if so - reject with that error; otherwise proceed expecting nothing but the content.

Note: Don't expect content to be Void - use Any? instead!